# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Enterprise untuk sistem audit log berbasis Blockchain dan Merkle Tree.",
        "title": "AuditChain Gateway API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@auditchain.local"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/report/generate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Men-generate laporan audit PDF atau CSV berdasarkan periode waktu tertentu.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/pdf",
                    "text/csv"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Generate audit report",
                "parameters": [
                    {
                        "description": "Parameter laporan",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/report.GenerateReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File laporan PDF atau CSV",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Format request tidak valid",
                        "schema": {
                            "$ref": "#/definitions/report.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Identitas client tidak valid",
                        "schema": {
                            "$ref": "#/definitions/report.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal generate laporan",
                        "schema": {
                            "$ref": "#/definitions/report.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "report.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Pesan kesalahan"
                }
            }
        },
        "report.GenerateReportRequest": {
            "type": "object",
            "required": [
                "format",
                "period_from",
                "period_to"
            ],
            "properties": {
                "format": {
                    "type": "string"
                },
                "period_from": {
                    "type": "string"
                },
                "period_to": {
                    "type": "string"
                },
                "sections": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}